glimpse-listener/requirements.txt
pyyaml==6.0.2
whisper
numpy
SpeechRecognition
torch
pydub
httpx